Split permissions_test into own suite
